% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{make_data}
\usage{
make_data(k1 = 3, alpha1 = 4, phi1 = 6, seed = NULL)
}
\arguments{
\item{k1}{mesor difference between groups.}

\item{alpha1}{amplitude difference between groups.}

\item{phi1}{Amount of phase shift, in hours, between the two groups within the generated data}

\item{seed}{random seed for generating data.}
}
\value{
data.frame
}
\description{
Generate example circadian data with specified phase shift between groups
}
\examples{
data <- make_data(k1=3, alpha1=4, phi1 = 6)
}
