% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{make_data}
\usage{
make_data(
  k = 0,
  k1 = 3,
  alpha = 10,
  alpha1 = 4,
  phi = 0,
  phi1 = 3.15,
  tau = 24,
  hours = 48,
  noise_sd = 0.1,
  seed = NULL
)
}
\arguments{
\item{k}{mesor of group 1.}

\item{k1}{change in mesor in group 2 from group 1.}

\item{alpha}{amplitude rhythm for group 1.}

\item{alpha1}{change in amplitude in group 2 from group 1}

\item{phi}{phase of rhythm, in radian-hours, in group 1.}

\item{phi1}{change in phase, in radian-hours, in group 2 from group 1}

\item{tau}{period of the rhythm, shared between both groups.}

\item{hours}{the number of hours/datapoints to sample.}

\item{noise_sd}{the standard deviation of the noise term.}

\item{seed}{random seed for generating data.}
}
\value{
data.frame
}
\description{
Generate example circadian data with specified phase shift between groups
}
\examples{
data <- make_data(k1 = 3, alpha1 = 4, phi1 = 6)
}
